Fix Dockerfile to point to dev; rewrite of app, but not working currently. Getting an error about Proxies
