+++ Dockerfile
@@ -1,8 +1,8 @@
 FROM python:3.11-slim
 
 WORKDIR /app
-COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
+COPY requirements.dev.txt .
+RUN pip install --no-cache-dir -r requirements.dev.txt
 
 COPY src/ src/
