+++ requirements.dev.txt
@@ -1,7 +1,7 @@
 flask==3.0.0
-Flask-Cors==4.0.0
-gunicorn==21.2.0
 anthropic==0.9.0
+PyPDF2==3.0.1
 python-dotenv==1.0.0
 Werkzeug==3.0.1
-PyPDF2==3.0.1
+Flask-Cors==4.0.0
+httpx==0.28.1
